+++ examples/cv64/scratches/func_8013B270_BE460/source.c
@@ -1,0 +1,11 @@
+void func_8013B270_BE460(Model* src, Model* dest) {
+    dest->position.x = src->position.x;
+    dest->position.y = src->position.y;
+    dest->position.z = src->position.z;
+    dest->size.x = src->size.x;
+    dest->size.y = src->size.y;
+    dest->size.z = src->size.z;
+    dest->angle.pitch = src->angle.pitch;
+    dest->angle.yaw = src->angle.yaw;
+    dest->angle.roll = src->angle.roll;
+}